app16_flask-form/requirements.txt
Flask==3.0.2
flask_mail==0.9.1
flask_sqlalchemy==3.1.1
python-dotenv==1.0.1
